+++ CHANGES.rst
@@ -5,10 +5,15 @@ Luftdatenpumpe changelog
 
 in progress
 ===========
+
+
+2019-05-21 0.14.0
+=================
 - Fix silly mixup with "is_active" indicator
 - Ignore ``--country=BE`` when operating on IRCELINE
 - Add new synthetic database-view fields
   "road_and_name_and_id" and "sos_feature_and_id"
+- Add new ``json.flex`` output target for flexible fieldname mapping
 
 
 2019-05-20 0.13.0